fix: removing hook for injecting global styles
