+++ .changeset/orange-buckets-rescue.md
@@ -1,0 +1,5 @@
+---
+'@project44-manifest/react-provider': patch
+---
+
+Useing cssbaseline component in provider
