+++ packages/react/components/provider/moon.yml
@@ -15,6 +15,8 @@ tasks:
 dependsOn:
   - id: 'button'
     scope: 'development'
+  - id: 'css-baseline'
+    scope: 'production'
   - id: 'styles'
     scope: 'production'
   - id: 'types'
