+++ packages/react/components/provider/package.json
@@ -34,6 +34,7 @@
     "react-dom": "^17.0.0 || ^18.0.0"
   },
   "dependencies": {
+    "@project44-manifest/react-css-baseline": "^0.1.0",
     "@project44-manifest/react-styles": "^1.3.0",
     "@project44-manifest/react-utils": "^0.3.0",
     "@react-aria/i18n": "^3.6.3",
