+++ packages/react/components/provider/src/Provider.styles.ts
@@ -1,3 +1,3 @@
 import { styled } from '@project44-manifest/react-styles';
 
-export const StyledProviderContent = styled('div', {});
+export const StyledProvider = styled('div', {});
