--- packages/react/components/provider/src/ProviderContent/index.ts
@@ -1,2 +1,0 @@
-export * from './ProviderContent';
-export * from './ProviderContent.types';
