+++ packages/react/components/provider/tsconfig.json
@@ -22,6 +22,9 @@
       "path": "../button"
     },
     {
+      "path": "../css-baseline"
+    },
+    {
       "path": "./tsconfig.build.json"
     }
   ]
